requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
breadability==0.1.20
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-icons==23.1
django-jazzmin==2.6.0
django-static-precompiler==2.4
docopt==0.6.2
ffmpeg==1.4
filelock==3.12.0
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
lit==16.0.3
llvmlite==0.40.0
lxml==4.9.2
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
mutagen==1.46.0
networkx==3.1
nltk==3.8.1
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
openai==0.27.6
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
pycountry==22.3.5
pycparser==2.21
pycryptodomex==3.17
PyJWT==2.6.0
PyPDF2==3.0.1
python-dotenv==1.0.0
python3-openid==3.2.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
sumy==0.11.0
sympy==1.11.1
tiktoken==0.3.3
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
websockets==11.0.3
whisper==1.1.10
yarl==1.9.2
yt-dlp==2023.3.4
